# docs/swagger.yaml
basePath: /
definitions:
  models.ChangeLog:
    properties:
      action:
        type: string
      created_at:
        type: string
      details:
        type: string
      id:
        type: integer
      task_id:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
  models.CreateTaskRequest:
    properties:
      description:
        type: string
      due_date:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      title:
        type: string
    required:
    - title
    type: object
...
  models.Task:
    properties:
      archived:
        type: boolean
      created_at:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/models.TaskStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.TaskStatus:
    enum:
    - To Do
    - In Progress
    - Done
    type: string
    x-enum-varnames:
    - StatusToDo
    - StatusInProgress
    - StatusDone
...
  models.UpdateTaskRequest:
    properties:
      description:
        type: string
      due_date:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    Task management system with authentication, comments, and change logs
    API for managing tasks, comments, and tracking changes with user authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Candidate Backend API
  version: "1.0"
paths:
...
  /api/tasks:
    get:
      consumes:
      - application/json
      description: Retrieve all non-archived tasks with creator information (supports
        pagination)
      parameters:
      - description: 'Limit number of results (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new task with title, description, and status
      parameters:
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new task
      tags:
      - Tasks
  /api/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task (only the creator can delete)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a task
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Retrieve a specific task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update task information (only the creator can update)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update a task
      tags:
      - Tasks
  /api/tasks/{id}/archive:
    post:
      consumes:
      - application/json
      description: Archive a task (only the creator can archive)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Archive a task
      tags:
      - Tasks
...
  /api/tasks/{id}/logs:
    get:
      consumes:
      - application/json
      description: Retrieve all change logs for a specific task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ChangeLog'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get task change logs
      tags:
      - Tasks
  /api/tasks/{id}/unarchive:
    post:
      consumes:
      - application/json
      description: Restore an archived task (only the creator can unarchive)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Unarchive a task
      tags:
      - Tasks
  /api/tasks/archived:
    get:
      consumes:
      - application/json
      description: Retrieve all archived tasks with creator information (supports
        pagination)
      parameters:
      - description: 'Limit number of results (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get archived tasks
      tags:
      - Tasks
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"